Partyman: always rescan on loading a track
I presume the idea is to use stored information about a track
and spare the work of re-scanning. But since you can start
dermixd with options that influence the length and possibly
the average loudness of tracks, rather always scan after loading
to have the proper view.

I am not yet fully sure that zeroscanning and Partyman interact
nicely. Need to verify with more testing that the track end is
handled and anticipated properly (setting crossfade point?).